requirements.txt
# Flask and related packages
Flask==2.3.2

Flask-Cors==3.0.10
psycopg2-binary==2.9.6


# Transformers and PyTorch for embeddings and LLM
transformers==4.32.0
torch==2.0.1
tqdm==4.66.1

# Pandas and NumPy for data manipulation
pandas==2.1.1
numpy==1.24.3

# For JSON handling
orjson==3.9.7

# Optional: Any other tools you might be using
gunicorn==20.1.0  # For deploying Flask in production (optional)
